fixed some issues
fixed the timing issues, now application adds randoms users to the user list during all instances of the list. (user in the list, user first in list, ect)

implimented more css changes to header, displays for pets and how the list is rendered in relation to the "add to list" button.

all functionality now syncs with the back end and  both front end and backend's are live.